requirements.txt
dj-database-url==0.5.0
Django==1.11.6
django-ckeditor==5.4.0
django-js-asset==1.0.0
django-render-block==0.5
django-templated-email==2.2.0
gunicorn==19.8.1
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
six==1.11.0
whitenoise==3.3.1
